Move iterator tests into a new file
