+++ frontend/test/resolution/CMakeLists.txt
@@ -46,6 +46,7 @@ comp_unit_test(testGetSymbolsAvailableInScope)
 comp_unit_test(testIf)
 comp_unit_test(testInitSemantics)
 comp_unit_test(testInteractive)
+comp_unit_test(testIterators)
 comp_unit_test(testLibrary)
 comp_unit_test(testLocation)
 comp_unit_test(testLoopIndexVars)
